fixes to improve parsing of JSON response
After April 1st 2023 I've started having some issues with the JSON returned by lnbits, the input was flagged as invalid. I've added this code to first select the data between {}, and then run the parser to get the JSON in the right format. 
I'm skipping the -- while (client.connected()) -- because in my code was not mandatory, but I acknowledge it could be a good practice to keep it.
Hope this fix helps!