Clarify how to fix compilation issues
Libraries are only checked into the libraries/ folder if specific or adapted versions
are necessary. Otherwise, they should be installed from the Arduino Library Manager.

But the names of the .h imports don't always correspond to the exact library names,
so this commit adds the exact names of the necessary libraries in a comment, right above
the line that will fail to compile if the libraries are missing.