Show small "arrow" in front of selected menu item
Real world "dummy" user inteface testing revealed that, when only 2 menu items are present
and they each have a different color (one selected, one not selected), the user doesn't
know which item is selected and which one is not. This results in them pressing "#SELECT"
on the wrong menu item.

Adding a small "arrow" in front of the select menu item fixes this, and it looks nice.