Rename "ATM" to more explicit "Send Offline"
This menu option is used to send Bitcoin over the Lightning network in an offline manner, using LNURL-withdraw.

Only US English speakers use the expression "ATM", as Canadians use "ABM", while the British
call them cashpoints or cash machines. Non-English speakers also use different names,
such as 'geldautomaat' in Flanders or 'pinautomaat' in the Netherlands.

The "Send Offline" functionality *can* be used to sell Bitcoin, not unlike Automated Teller Machine,
although in this case it's the opposite of automated, as it's very much human operated.

But this functionality might just as well be used to withdraw funds from the device, or to refund a customer.

Therefore, it's more clear and in line with the other menu items to rename "ATM" to simply "Send Offline".